+++ src/common/bizHooks/index.js
@@ -1,1 +1,3 @@
 export {useParams} from './useParams'
+export {useQuery} from './useQuery'
+export {useSelectNodes} from './useSelectNodes'